requirements.txt
matplotlib
imageio
h5py
numpy==1.15.4
Pillow==7.0.0
six==1.14.0
torch==1.4.0
torchvision==0.5.0
